@@ -8,8 +8,8 @@
 	"dependencies": {
 		"@material-ui/core": "^4.11.0",
 		"@material-ui/icons": "^4.9.1",
-		"react": "^17.0.1",
-		"react-dom": "^17.0.1",
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0",
 		"react-router-dom": "^5.2.0"
 	},
 	"devDependencies": {
